# api/openapi/service.yaml
openapi: 3.0.3
info:
  title: billing-service
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://billing-service.internal.gonkagate.local
    description: placeholder internal service base URL
security: []
...
paths:
...
  /internal/billing/v1/accounts/resolve:
    post:
      tags: [billing]
      summary: Resolve represented caller identity to a billing account
      operationId: resolveBillingAccount
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.accounts.resolve]
      x-security-decision:
        exposure: protected
        rationale: service-to-service account binding and migration-readiness path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountResolveRequest"
      responses:
        "200":
          description: account resolve outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResolveResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /internal/billing/v1/usage/reservations:
    post:
      tags: [billing]
      summary: Create or read back migrated usage reserve lineage
      operationId: reserveUsage
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.write]
      x-security-decision:
        exposure: protected
        rationale: money-lineage write path for usage reserve under billing authority
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageReserveRequest"
      responses:
        "200":
          description: usage reserve outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/finalizations:
    post:
      tags: [billing]
      summary: Finalize a usage operation within microlease authority
      operationId: finalizeUsage
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.write]
      x-security-decision:
        exposure: protected
        rationale: money-effect terminal write path for usage finalization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageTerminalRequest"
      responses:
        "200":
          description: usage finalization outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/write-offs:
    post:
      tags: [billing]
      summary: Write off or release a usage operation explicitly
      operationId: writeOffUsage
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.write]
      x-security-decision:
        exposure: protected
        rationale: explicit support-safe write-off path for unchargeable usage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageTerminalRequest"
      responses:
        "200":
          description: usage write-off outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/reversals:
    post:
      tags: [billing]
      summary: Reverse a previously committed usage effect
      operationId: reverseUsage
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.write]
      x-security-decision:
        exposure: protected
        rationale: explicit compensation path for committed usage effects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageReversalRequest"
      responses:
        "200":
          description: usage reversal outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/usage/readback:
    post:
      tags: [billing]
      summary: Read usage operation state by durable usage identity
      operationId: readUsageOperation
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.usage.read]
      x-security-decision:
        exposure: protected
        rationale: account-bound usage readback path for ambiguous outcomes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsageReadbackRequest"
      responses:
        "200":
          description: usage readback outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageOperationReadbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/microleases/issue:
    post:
      tags: [billing]
      summary: Issue or replenish a billing microlease
      operationId: issueMicrolease
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.microleases.write]
      x-security-decision:
        exposure: protected
        rationale: money-affecting service-to-service microlease issuance path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MicroleaseIssueRequest"
      responses:
        "200":
          description: issue or replay outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroleaseIssueResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/microleases/readback:
    post:
      tags: [billing]
      summary: Read a microlease or stored operation outcome
      operationId: readMicrolease
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.microleases.read]
      x-security-decision:
        exposure: protected
        rationale: support-safe service-to-service microlease readback path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MicroleaseReadbackRequest"
      responses:
        "200":
          description: readback outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroleaseReadbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/microleases/close:
    post:
      tags: [billing]
      summary: Close or checkpoint a microlease with allocator proof
      operationId: closeMicrolease
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.microleases.write]
      x-security-decision:
        exposure: protected
        rationale: money-affecting service-to-service close and release path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MicroleaseCloseRequest"
      responses:
        "200":
          description: close or checkpoint outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroleaseCloseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "423":
          $ref: "#/components/responses/Locked"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/operations/readback:
    post:
      tags: [billing]
      summary: Read a stored operation outcome
      operationId: readBillingOperation
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.operations.read]
      x-security-decision:
        exposure: protected
        rationale: support-safe service-to-service operation readback path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingOperationReadbackRequest"
      responses:
        "200":
          description: operation readback outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingOperationReadbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /internal/billing/v1/reconciliation/cases:
    get:
      tags: [billing]
      summary: List support-safe reconciliation cases
      operationId: listReconciliationCases
      security:
        - ServiceBearerAuth: []
      x-route-scopes: [billing.reconciliation.read]
      x-security-decision:
        exposure: protected
        rationale: scoped support readback for reconciliation cases
      parameters:
        - name: accountScopeKey
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [open, leased, waiting_evidence, manual_review, resolved, canceled]
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, critical]
      responses:
        "200":
          description: reconciliation cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationCasesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    ServiceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Platform-issued service identity token for protected internal billing routes.
  schemas:
...
    ContractVersion:
      type: string
      enum: [v1]
...
    AccountResolveRequest:
      type: object
      required:
        - routeContractId
        - contractVersion
        - callerContext
        - representedUserContext
        - deadlineAtEpochMs
        - traceRequestId
      additionalProperties: false
      properties:
        routeContractId:
          type: string
          enum: [billing.accounts.resolve]
        contractVersion:
          $ref: "#/components/schemas/ContractVersion"
        callerContext:
          $ref: "#/components/schemas/CallerContext"
        representedUserContext:
          $ref: "#/components/schemas/RepresentedUserContext"
        deadlineAtEpochMs:
          type: integer
          format: int64
          minimum: 0
        traceRequestId:
          type: string
          minLength: 1
    AccountResolveResponse:
      type: object
      required:
        - resultCode
        - accountScopeKey
        - accountState
        - importState
        - migrationState
        - balanceReadEligible
        - retryable
      additionalProperties: false
      properties:
        resultCode:
          $ref: "#/components/schemas/AccountResolveResultCode"
        accountScopeKey:
          type: string
          minLength: 1
        billingAccountId:
          type: string
          minLength: 1
        accountState:
          $ref: "#/components/schemas/AccountState"
        importState:
          $ref: "#/components/schemas/ImportState"
        migrationState:
          type: string
          enum: [legacy, shadow, internal_cohort, migrated, rollback]
        balanceReadEligible:
          type: boolean
        retryable:
          type: boolean
        failureClass:
          type: string
          minLength: 1
        balanceVersion:
          type: integer
          format: int64
          minimum: 0
...
    UsageReserveRequest:
      type: object
      required:
        - routeContractId
        - contractVersion
        - idempotencyKey
        - requestFingerprint
        - callerContext
        - representedUserContext
        - accountScopeKey
        - usageOperationId
        - authorityMode
        - microleaseId
        - microleaseChildDebitId
        - debitAuthorizationId
        - proxyAllocatorOwnerId
        - microleaseGeneration
        - leaseFence
        - childSequence
        - childCapUsdAtoms
        - pricingSnapshot
        - deadlineAtEpochMs
        - traceRequestId
      additionalProperties: false
      properties:
        routeContractId:
          type: string
          enum: [billing.usage.reserve]
        contractVersion:
          $ref: "#/components/schemas/ContractVersion"
        idempotencyKey:
          type: string
          minLength: 1
        requestFingerprint:
          type: string
          minLength: 1
        callerContext:
          $ref: "#/components/schemas/CallerContext"
        representedUserContext:
          $ref: "#/components/schemas/RepresentedUserContext"
        accountScopeKey:
          type: string
          minLength: 1
        usageOperationId:
          type: string
          minLength: 1
        authorityMode:
          $ref: "#/components/schemas/UsageAuthorityMode"
        microleaseId:
          type: string
          minLength: 1
        microleaseChildDebitId:
          type: string
          minLength: 1
        debitAuthorizationId:
          type: string
          minLength: 1
        proxyAllocatorOwnerId:
          type: string
          minLength: 1
        microleaseGeneration:
          type: integer
          format: int64
          minimum: 1
        leaseFence:
          type: string
          minLength: 1
        childSequence:
          type: integer
          format: int64
          minimum: 1
        childCapUsdAtoms:
          type: integer
          format: int64
          minimum: 1
        pricingSnapshot:
          $ref: "#/components/schemas/PricingSnapshotEvidence"
        deadlineAtEpochMs:
          type: integer
          format: int64
          minimum: 0
        traceRequestId:
          type: string
          minLength: 1
        safeOperationMetadata:
          $ref: "#/components/schemas/SafeOperationMetadata"
    UsageTerminalRequest:
      type: object
      required:
        - routeContractId
        - contractVersion
        - idempotencyKey
        - requestFingerprint
        - callerContext
        - accountScopeKey
        - usageOperationId
        - microleaseId
        - microleaseChildDebitId
        - debitAuthorizationId
        - terminalOutcomeId
        - terminalFingerprint
        - chargedUsdAtoms
        - releasedUsdAtoms
        - writeOffUsdAtoms
        - pricingSnapshot
        - deadlineAtEpochMs
        - traceRequestId
      additionalProperties: false
      properties:
        routeContractId:
          type: string
          enum: [billing.usage.finalize, billing.usage.write_off]
        contractVersion:
          $ref: "#/components/schemas/ContractVersion"
        idempotencyKey:
          type: string
          minLength: 1
        requestFingerprint:
          type: string
          minLength: 1
        callerContext:
          $ref: "#/components/schemas/CallerContext"
        accountScopeKey:
          type: string
          minLength: 1
        usageOperationId:
          type: string
          minLength: 1
        microleaseId:
          type: string
          minLength: 1
        microleaseChildDebitId:
          type: string
          minLength: 1
        debitAuthorizationId:
          type: string
          minLength: 1
        terminalOutcomeId:
          type: string
          minLength: 1
        terminalFingerprint:
          type: string
          minLength: 1
        chargedUsdAtoms:
          type: integer
          format: int64
          minimum: 0
        releasedUsdAtoms:
          type: integer
          format: int64
          minimum: 0
        writeOffUsdAtoms:
          type: integer
          format: int64
          minimum: 0
        pricingSnapshot:
          $ref: "#/components/schemas/PricingSnapshotEvidence"
        qualifiedInferenceEvidenceId:
          type: string
          minLength: 1
        deadlineAtEpochMs:
          type: integer
          format: int64
          minimum: 0
        traceRequestId:
          type: string
          minLength: 1
        safeOperationMetadata:
          $ref: "#/components/schemas/SafeOperationMetadata"
    UsageReversalRequest:
      type: object
      required:
        - routeContractId
        - contractVersion
        - idempotencyKey
        - requestFingerprint
        - callerContext
        - accountScopeKey
        - usageOperationId
        - originalLedgerEntryId
        - reversalUsdAtoms
        - reasonCode
        - deadlineAtEpochMs
        - traceRequestId
      additionalProperties: false
      properties:
        routeContractId:
          type: string
          enum: [billing.usage.reversal]
        contractVersion:
          $ref: "#/components/schemas/ContractVersion"
        idempotencyKey:
          type: string
          minLength: 1
        requestFingerprint:
          type: string
          minLength: 1
        callerContext:
          $ref: "#/components/schemas/CallerContext"
        accountScopeKey:
          type: string
          minLength: 1
        usageOperationId:
          type: string
          minLength: 1
        originalLedgerEntryId:
          type: string
          minLength: 1
        reversalUsdAtoms:
          type: integer
          format: int64
          minimum: 1
        reasonCode:
          type: string
          minLength: 1
        deadlineAtEpochMs:
          type: integer
          format: int64
          minimum: 0
        traceRequestId:
          type: string
          minLength: 1
        safeOperationMetadata:
          $ref: "#/components/schemas/SafeOperationMetadata"
...
    UsageOperationResponse:
      type: object
      required:
        - resultCode
        - usageOperationId
        - state
        - replayIdentity
      additionalProperties: false
      properties:
        resultCode:
          $ref: "#/components/schemas/UsageOperationResultCode"
        usageOperationId:
          type: string
          minLength: 1
        state:
          $ref: "#/components/schemas/UsageOperationState"
        replayIdentity:
          $ref: "#/components/schemas/ReplayIdentity"
        billingOperationId:
          type: string
          minLength: 1
        reconciliationCaseId:
          type: string
          minLength: 1
        reasonCode:
          type: string
          minLength: 1
    UsageOperationReadbackResponse:
      type: object
      required:
        - resultCode
        - usageOperationId
        - state
      additionalProperties: false
      properties:
        resultCode:
          $ref: "#/components/schemas/UsageOperationResultCode"
        usageOperationId:
          type: string
          minLength: 1
        state:
          $ref: "#/components/schemas/UsageOperationState"
        replayIdentity:
          $ref: "#/components/schemas/ReplayIdentity"
        reconciliationCaseId:
          type: string
          minLength: 1
        reasonCode:
          type: string
          minLength: 1
...
  responses:
    BadRequest:
      description: request is malformed or invalid
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: missing or invalid service authentication
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: authenticated caller lacks route scope or account binding
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: payload or state conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: business rejection such as insufficient funds or stale pricing
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Locked:
      description: manual review, suspended account, or locked state
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: admission throttle
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ServiceUnavailable:
      description: dependency or admission control is not ready
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalServerError:
      description: unexpected server failure
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"